MiniProjects/AI_Blog_Generation/pyproject.toml
[project]
name = "ai-blog-generation"
version = "0.1.0"
description = "AI Blog Generation with LangGraph - Generate blogs from topics with optional translation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "fastapi>=0.115.12",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "pydantic>=2.12.5",
    "uvicorn>=0.34.3",
    "watchdog>=6.0.0",
    "langgraph>=1.0.5",
    "langgraph-cli[inmem]>=0.3.1",
]
